Скорректирован код, добавлены проверки на наличие/отсутствие папко и файлов

+++ Program.cs
@@ -10,9 +10,10 @@ namespace FileManager
     class Program
     {
         static void Main(string[] args)
-        {
+        {            
             string command = Console.ReadLine(); //вводим команду с клавиатуры
             Console.Clear();
+
             if (command == "ls") //вывод файлов и каталогов в консоль
             {
                 Output();
@@ -21,7 +22,7 @@ namespace FileManager
             {
                 CopyDir();
             }
-            else if (command == "rm") //удаление каталога
+            if (command == "rm") //удаление каталога
             {
                 RemoveDir();
             }
@@ -37,58 +38,84 @@ namespace FileManager
             {
                 RemoveFile();
             }
-            else if (command == "file info") // вывод информации о файле - атрибуты и размер
+            else if(command == "file info") // вывод информации о файле - атрибуты и размер
             {
                 InfoFile();
             }
+
             Console.ReadLine();
         }
         static void Output()
         {
-            string address = Console.ReadLine(); //вводим адрес каталога            
-            string[] list = Directory.GetFileSystemEntries(address, "*", SearchOption.AllDirectories); //создаем массив из каталогов
-            int level = int.Parse(Console.ReadLine()); //указать номер страницы, которую необходимо открыть
-            Console.Clear(); //чистим консоль
-            Console.WriteLine("*************************************"); //выводим рамку на консоль
-            Console.WriteLine(address); //выводим адрес каталога
-            Console.WriteLine("*************************************"); //выводим рамку на консоль
-            Console.WriteLine("Page "+level); //Выводим номер страницы
+            string address = Console.ReadLine(); //вводим адрес каталога
+            bool correctAddress = Directory.Exists(address); //проверяем на наличие заданного каталога в компьютере
 
-            if (level == 1) 
+            if (correctAddress==true)
             {
-                for (int i = 0; i < 10; i++)
+                string[] list = Directory.GetFileSystemEntries(address, "*", SearchOption.AllDirectories); //создаем массив из каталогов
+                int level = int.Parse(Console.ReadLine()); //указать номер страницы, которую необходимо открыть
+                Console.Clear(); //чистим консоль
+                Console.WriteLine("*************************************"); //выводим рамку на консоль
+                Console.WriteLine(address); //выводим адрес каталога
+                Console.WriteLine("*************************************"); //выводим рамку на консоль
+                Console.WriteLine("Page "+level); //Выводим номер страницы
+
+                if (level == 1) 
                 {
-                    Console.WriteLine(list[i]); //выводим файлы и каталоги 1 страницы
+                    for (int i = 0; i < 10; i++)
+                    {
+                        Console.WriteLine(list[i]); //выводим файлы и каталоги 1 страницы
+                    }
+                }
+                else if (level == 2)
+                {
+                    for (int i = 10; i < list.Length; i++)
+                    {
+                        Console.WriteLine(list[i]); //выводим файлы и каталоги 2 страницы
+                    }
                 }
             }
-            else if (level == 2)
+            else
             {
-                for (int i = 10; i < list.Length; i++)
-                {
-                    Console.WriteLine(list[i]); //выводим файлы и каталоги 2 страницы
-                }
+                Console.WriteLine("Такого каталога не существует!");
             }
+
         }
         static void CopyDir()
         {
             string address = Console.ReadLine(); //Вводим адрес каталога, который хотим скопировать
-            Console.WriteLine(Directory.Exists(address)); //проверяем на наличие заданного каталога в компьютере
-            Console.Clear(); //чистим консоль
-            string targetPath = Console.ReadLine(); //Вводим адрес папки, куда мы хотим её скопировать
+            bool correctAddress = Directory.Exists(address); //проверяем на наличие заданного каталога в компьютере
 
-            Directory.CreateDirectory(targetPath); //создаем по новому адресу "скопированную" папку
+            if (correctAddress)
+            {
+                Console.Clear(); //чистим консоль
+                string targetPath = Console.ReadLine(); //Вводим адрес папки, куда мы хотим её скопировать
+
+                Directory.CreateDirectory(targetPath); //создаем по новому адресу "скопированную" папку
+            }
+            else
+            {
+                Console.WriteLine("Такого каталога не существует!");
+            }
         }
         static void RemoveDir()
         {
             string address = Console.ReadLine(); //вводим адрес каталога 
-            Console.WriteLine(Directory.Exists(address)); //проверяем на наличие заданного каталога в компьютере
-            Directory.Delete(address); //удалаем каталог по заданному адресу
-            Console.Clear(); //чистим консоль
+            bool correctAddress = Directory.Exists(address); //проверяем на наличие заданного каталога в компьютере
+            if (correctAddress)
+            {
+                Directory.Delete(address); //удалаем каталог по заданному адресу
+                Console.Clear(); //чистим консоль
+            }
+            else
+            {
+                Console.WriteLine("Такого каталога не существует!");
+            }           
+
         }
         static void InfoDir()
         {
             string address = Console.ReadLine(); //вводим адрес каталога
-            Console.WriteLine(Directory.Exists(address)); //проверяем на наличие заданного каталога в компьютере
             Console.Clear(); //чистим консоль
 
             Output(); //выводим содержимое каталога на консоль
@@ -102,30 +129,53 @@ namespace FileManager
         static void CopyFile()
         {
             string fileName = "test.txt"; //имя файла для копирования
-            string sourcePath = Console.ReadLine(); //вводим адрес файла для копирования
-            string targetPath = Console.ReadLine(); //вводим новый адрес, куда файл будет скопирован
+            string sourcePath = Console.ReadLine(); //вводим адрес файла для копирования с указанием имени и расширения файла
+            bool correctFilename = File.Exists(sourcePath);
 
-            string sourceFile = Path.Combine(sourcePath, fileName); //Указываем адрес исходного файла
-            string destFile = Path.Combine(targetPath, fileName); //Указываем адрес полученного файла 
-
-            File.Copy(sourceFile, destFile, true); //копируем файл 
+            if (correctFilename)
+            {                
+                string targetPath = Console.ReadLine(); //вводим новый адрес, куда файл будет скопирован
+                string sourceFile = Path.Combine(sourcePath, fileName); //Указываем адрес исходного файла
+                string destFile = Path.Combine(targetPath, fileName); //Указываем адрес полученного файла 
+                File.Copy(sourceFile, destFile, true); //копируем файл 
+            }
+            else
+            {
+                Console.WriteLine("Такого файла не существует!");
+            }
         }
         static void RemoveFile()
         {
             string address = Console.ReadLine(); //вводим адрес файла для удаления
-            Console.WriteLine(File.Exists(address)); //проверяем на наличие заданного файла в компьютере
-            File.Delete(address); // удаляем файл по определенному адресу
-            Console.Clear(); //чистим консоль
+            bool correctAddress = File.Exists(address); //проверяем на наличие заданного файла в компьютере
+
+            if (correctAddress == true)
+            {
+                File.Delete(address); // удаляем файл по определенному адресу
+                Console.Clear(); //чистим консоль
+            }
+            else
+            {
+                Console.WriteLine("Такого файла не существует!");
+            }
         }
         static void InfoFile()
         {
             string address = Console.ReadLine(); //вводим адрес файла
-            Console.WriteLine(File.Exists(address)); //проверяем на наличие заданного файла в компьютере
-            FileInfo fileInfo = new FileInfo(address); //создаем класс
-            Console.WriteLine("*************************************");
-            Console.WriteLine(fileInfo.Attributes.ToString()); //выводит атрибуты файла
-            Console.WriteLine(fileInfo.Length.ToString()); //выводит размер файла в байтах
-            Console.WriteLine("*************************************");
+            bool correctAddress = File.Exists(address); //проверяем на наличие заданного файла в компьютере
+
+            if(correctAddress==true) 
+            {
+                FileInfo fileInfo = new FileInfo(address); //создаем класс
+                Console.WriteLine("*************************************");
+                Console.WriteLine(fileInfo.Attributes.ToString()); //выводит атрибуты файла
+                Console.WriteLine(fileInfo.Length.ToString()); //выводит размер файла в байтах
+                Console.WriteLine("*************************************");
+            }
+            else
+            {
+                Console.WriteLine("Такого файла не существует!");
+            }     
         }
     }
 }
